Modify number type to be Float

+++ app/src/main/java/com/example/jetpack_compose_calculator/Globals.kt
@@ -14,7 +14,7 @@ var operatorType: String by mutableStateOf("")
 var operatorClicked: Boolean by mutableStateOf(false)
 
 // this variable will be used to store the numbers that will be used to perform any math operation
-var numbers = mutableListOf<Int>()
+var numbers = mutableListOf<Float>()
 
 val operatorsFunctionality = listOf(
     MathOperator(Icons.Outlined.Add, "Adición", {
